Prepare kin-infer 0.2.10 C7 attention groundwork
Merge groundwork: gated resident diagnostics, transient norm-buffer fix, default-off C7 plan/gate skeleton, streaming attention oracle tests, clippy cleanup, and kin-infer 0.2.10 version bump.

+++ tests/metal_oom_error_path.rs
@@ -51,7 +51,7 @@ use metal::Device;
 fn max_buffer_length() -> u64 {
     Device::system_default()
         .expect("a system-default Metal device")
-        .max_buffer_length() as u64
+        .max_buffer_length()
 }
 
 /// A count of f32 elements whose byte footprint strictly exceeds the device's
